+++ src/clientlist.c
@@ -70,7 +70,7 @@ void remove_from_list(DLink **list, void *ptr, DLink *link)
       }
    }
 
-   sendto_realops("remove_from_list(%x, %x) failed!!", (int) list, (int) ptr);
+   sendto_realops("remove_from_list(%x, %x) failed!!", (long) list, (long) ptr);
 }
 
 u_long
